e2e

+++ egs/swbd/s5c/local/chain/adaptation/LHUC/BLHUC_adaptation.sh
@@ -4,6 +4,7 @@ set -e
 
 # configs for 'chain'
 stage=0
+stop_stage=100
 train_stage=-10
 get_egs_stage=-10
 baseline=cnn_tdnn_iv_1a_hires_sp
@@ -12,6 +13,8 @@ test_ivector_dir=exp/nnet3/ivectors_eval2000
 epoch_num=7
 lr1=0.01
 lr2=0.01
+num_jobs_init=1
+num_jobs_final=1
 num_chunk=64
 param_mean_init=0.0
 param_std_init=1.0
@@ -20,25 +23,57 @@ log_std=false
 tied_std=true
 std_update_scale=6.51e-4 # give a small scale if tied_std == true
 tag= # any other marks
+pre_out_dim=256 # 384
+pre_out_layer=prefinal-l # lstm3.rp
+output_chain_layer=prefinal-chain.affine # output.affine
+
+alignment_subsampling_factor=1
+constrain=false
 
 decode_iter=
 decode_nj=50
+decode_opt=
 
 # training options
 frames_per_eg=150,100,50,20,10,5
 remove_egs=false
 xent_regularize=1.0
+add_opt="--optimization.memory-compression-level=2"
+mmi_scale=0.0
+l2_regularize=0
+chunk_left_context=
+chunk_right_context=
+chunk_left_context_initial=
+chunk_right_context_final=
+deriv_truncate_margin=
+frames_per_chunk_primary=
+extra_left_context=
+extra_right_context=
+extra_left_context_initial=
+extra_right_context_final=
 
 adapted_layer="cnn1 tdnnf7 tdnnf8 tdnnf9 tdnnf10 tdnnf11 tdnnf12"
 layer_dim="2560 1536 1536 1536 1536 1536 1536" # should be corresponding to the $adapted_layer
 KL_scale="0.0001 1.0 1.0 1.0 1.0 1.0 1.0"
 input_config="component-node name=idct component=idct input=feature1" # cnn-tdnn, for tdnn, it can be "component-node name=lda component=lda input=Append(Offset(feature1, -1), feature1, Offset(feature1, 1), ReplaceIndex(ivector, t, 0))"
+input_config1=
+input_config2=
+input_config3=
 input_dim=41
 prior_mean_file=  # a file or a list of files
 prior_std_file=  # a file or a list of files
 prior_mean="0.0 0.0" # a value or a list of values
 prior_std="1.0 1.0"  # a value or a list of values
 common_egs_dir=
+deriv_weights_scp=
+spk_count_file=
+adapt_base=adaptation/LHUC
+adapt_model=
+init_model=
+latbase=
+LM_path=data/lang_
+
+cuda_id="0,1,2,3,4,5,6,7,8"
 
 # End configuration section.
 echo "$0 $@"  # Print the command line for logging
@@ -46,6 +81,8 @@ echo "$0 $@"  # Print the command line for logging
 . ./cmd.sh
 . ./path.sh
 . ./utils/parse_options.sh
+
+export CUDA_VISIBLE_DEVICES=$cuda_id
 
 adapted_layer_array=($adapted_layer)
 layer_dim_array=($layer_dim)
@@ -64,20 +101,32 @@ adapt_set=$1 # eval2000_hires_spk_sub20
 label_lat_dir=$2 # label_lat_dir=exp/chain/cnn_tdnn_iv_1a_hires_sp/decode_eval2000_hires_sw1_fsh_fg/1BEST_lat/score_10_0.0
 decode_set=$3 # eval2000_hires_spk
 
-version=_BLHUC${tag}_adaptlayer${layer_num}_act${act}_batch${num_chunk}_epoch${epoch_num}_lr1${lr1}_lr2${lr2}
+num_chunk1=`echo $num_chunk | tr "," "_"`
+version=_BLHUC${tag}_adaptlayer${layer_num}_act${act}_batch${num_chunk1}_epoch${epoch_num}_lr1${lr1}_lr2${lr2}
 
 dirbase=exp/chain/${baseline}
-dir=exp/chain/adaptation/LHUC/${baseline}${version}
+dir=exp/chain/$adapt_base/${baseline}${version}
 
-if [ $stage -le 0 ]; then
 
 mkdir -p $dir
 cp -r $dirbase/{configs,phones.txt,phone_lm.fst,tree,den.fst,normalization.fst,0.trans_mdl} $dir/
-cp -r $dirbase/{final.mdl,tree,phones.txt} $label_lat_dir/
+
+if [ -z $latbase ]; then
+    cp -r $dirbase/{tree,phones.txt} $label_lat_dir/
+else
+    cp -r $latbase/{tree,phones.txt} $label_lat_dir/
+fi
+
+
+if [ $stage -le 0 ] && [ $stop_stage -ge 0 ]; then
 
 spk_count=data/${adapt_set}/spk_count
 spk_num=`cat data/${adapt_set}/num_spk`
 input_dim_nospk=$(awk "BEGIN{print($input_dim-1)}")
+
+if [ ! -z $spk_count_file ]; then
+    spk_count=$spk_count_file
+fi
 
 # input features
 cat <<EOF > $dir/configs/change.config
@@ -86,9 +135,12 @@ cat <<EOF > $dir/configs/change.config
 	# acoustic features
 	dim-range-node name=feature1 input-node=input dim=$input_dim_nospk dim-offset=0
 	${input_config}
+    ${input_config1}
+    ${input_config2}
+    ${input_config3}
 
 	# speaker id
-	dim-range-node name=feature2 input-node=input dim=1 dim-offset=40
+	dim-range-node name=feature2 input-node=input dim=1 dim-offset=$input_dim_nospk
 	
 	# frame count per speaker
 	component name=BLHUC.count type=LinearSelectColComponent input-dim=1 output-dim=1 col-num=$spk_num matrix=$spk_count max-change=0 learning-rate-factor=0 l2-regularize=0.00 use-natural-gradient=false
@@ -149,7 +201,7 @@ if [ ! -z $prior_std_file ]; then
   prior_std_config="vector=$prior_std_file_tmp"
 fi
 
-if [[ "$i" < "${#prior_mean_array[*]}" || "$i" < "${#prior_mean_file_array[*]}" || "$i" < "${#prior_std_array[*]}" || "$i" < "${#prior_std_file_array[*]}" ]]; then
+if [[ "$i" -lt "${#prior_mean_array[*]}" || "$i" -lt "${#prior_mean_file_array[*]}" || "$i" -lt "${#prior_std_array[*]}" || "$i" -lt "${#prior_std_file_array[*]}" ]]; then
   prior_layer=$layer
 cat <<EOF >> $dir/configs/change.config
 	component name=BLHUC.prior_mean.$prior_layer type=ConstantFunctionComponent input-dim=1 is-updatable=false output-dim=$dim_tmp $prior_mean_config
@@ -205,42 +257,69 @@ done
 
 # use cross entropy only
 cat <<EOF >> $dir/configs/change.config
-	component name=no_mmi type=NoOpComponent dim=256 backprop-scale=0.0
-	component-node name=no_mmi component=no_mmi input=prefinal-l
-	component-node name=prefinal-chain.affine component=prefinal-chain.affine input=no_mmi
+	component name=no_mmi type=NoOpComponent dim=$pre_out_dim backprop-scale=$mmi_scale
+	component-node name=no_mmi component=no_mmi input=${pre_out_layer}
+	component-node name=$output_chain_layer component=$output_chain_layer input=no_mmi
 EOF
 
-nnet3-am-copy --raw --binary=false --edits="set-learning-rate-factor learning-rate-factor=0" $dirbase/final.mdl - | \
+if [ -z $adapt_model ]; then
+    adapt_model=$dirbase/final.mdl
+fi
+
+if [ -z $latbase ]; then
+    cp -r $adapt_model $label_lat_dir/
+else
+    cp -r $latbase/final.mdl $label_lat_dir/
+fi
+
+nnet3-am-copy --raw --binary=false --edits="set-learning-rate-factor learning-rate-factor=0" $adapt_model - | \
  sed "s/<TestMode> F/<TestMode> T/g" | sed "s/BatchNormComponent/BatchNormTestComponent/g" | sed "s/<OrthonormalConstraint> [^ ]* /<OrthonormalConstraint> 0/g" | \
- nnet3-copy --nnet-config=$dir/configs/change.config - $dir/0.raw
+ nnet3-copy --nnet-config=$dir/configs/change.config - - | nnet3-copy --edits="remove-orphans" - $dir/0.raw
 
 nnet3-info $dir/0.raw > $dir/0.raw.info
 
 fi
 
-if [ $stage -le 1 ]; then
-  local/chain/adaptation/train_adapt.py --stage $train_stage \
+if [ ! -z $init_model ]; then
+    nnet3-copy $init_model $dir/0.raw
+    nnet3-info $dir/0.raw > $dir/0.raw.info
+fi
+
+if [ $stage -le 1 ] && [ $stop_stage -ge 1 ]; then
+  
+  apply_deriv_weights=false
+  if [ ! -z $deriv_weights_scp ]; then
+    apply_deriv_weights=true
+  fi
+  
+  steps/nnet3/chain/train_adapt.py --stage $train_stage \
     --cmd "$train_cmd" \
-	--feat.online-ivector-dir $adapt_ivector_dir \
+    --use-gpu "wait" ${deriv_weights_scp:+ --egs.deriv-weights-scp $deriv_weights_scp} \
+	${adapt_ivector_dir:+ --feat.online-ivector-dir $adapt_ivector_dir} \
     --feat.cmvn-opts "--norm-means=false --norm-vars=false" \
     --chain.xent-regularize $xent_regularize \
     --chain.leaky-hmm-coefficient 0.1 \
-    --chain.l2-regularize 0.0 \
-    --chain.apply-deriv-weights false \
+    --chain.l2-regularize $l2_regularize \
+    --chain.apply-deriv-weights $apply_deriv_weights \
     --chain.lm-opts="--num-extra-lm-states=2000" \
-	--chain.alignment-subsampling-factor 1 \
-    --trainer.add-option="--optimization.memory-compression-level=2" \
+	--chain.alignment-subsampling-factor $alignment_subsampling_factor \
+    --trainer.add-option="$add_opt" \
     --egs.dir "$common_egs_dir" \
     --egs.stage $get_egs_stage \
-    --egs.opts "--frames-overlap-per-eg 0 --constrained false" \
+    --egs.opts "--frames-overlap-per-eg 0 --constrained $constrain" \
     --egs.chunk-width $frames_per_eg \
+    ${chunk_left_context:+ --egs.chunk-left-context $chunk_left_context} \
+    ${chunk_right_context:+ --egs.chunk-right-context $chunk_right_context} \
+    ${chunk_left_context_initial:+ --egs.chunk-left-context-initial $chunk_left_context_initial} \
+    ${chunk_right_context_final:+ --egs.chunk-right-context-final $chunk_right_context_final} \
     --trainer.num-chunk-per-minibatch ${num_chunk} \
     --trainer.frames-per-iter 1500000 \
     --trainer.num-epochs $epoch_num \
-    --trainer.optimization.num-jobs-initial 1 \
-    --trainer.optimization.num-jobs-final 1 \
+    --trainer.optimization.num-jobs-initial $num_jobs_init \
+    --trainer.optimization.num-jobs-final $num_jobs_final \
     --trainer.optimization.initial-effective-lrate $lr1 \
     --trainer.optimization.final-effective-lrate $lr2 \
+    ${deriv_truncate_margin:+ --trainer.deriv-truncate-margin $deriv_truncate_margin} \
 	--trainer.optimization.do-final-combination false \
     --trainer.max-param-change 2.0 \
 	--trainer.input-model $dir/0.raw \
@@ -251,16 +330,17 @@ if [ $stage -le 1 ]; then
 
   mv $dir/final.mdl $dir/final_ori.mdl
   nnet3-am-copy --binary=false $dir/final_ori.mdl - | \
-   sed "s/<TestMode> F/<TestMode> T/g" > $dir/final.mdl
+   sed "s/<TestMode> F/<TestMode> T/g" | \
+   nnet3-am-copy --binary=true --edits="remove-orphans" - $dir/final.mdl
 
   nnet3-am-info $dir/final.mdl > $dir/final.mdl.info
 fi
 
-if [ $stage -le 2 ]; then
+if [ $stage -le 2 ] && [ $stop_stage -ge 2 ]; then
   # Note: it might appear that this $lang directory is mismatched, and it is as
   # far as the 'topo' is concerned, but this script doesn't read the 'topo' from
   # the lang directory.
-  utils/mkgraph.sh --self-loop-scale 1.0 data/lang_sw1_tg $dir $dir/graph_sw1_tg
+  utils/mkgraph.sh --self-loop-scale 1.0 ${LM_path}sw1_tg $dir $dir/graph_sw1_tg
 fi
 
 
@@ -269,17 +349,22 @@ iter_opts=
 if [ ! -z $decode_iter ]; then
   iter_opts=" --iter $decode_iter "
 fi
-if [ $stage -le 3 ]; then
+if [ $stage -le 3 ] && [ $stop_stage -ge 3 ]; then
   rm $dir/.error 2>/dev/null || true
   (
   steps/nnet3/decode.sh --acwt 1.0 --post-decode-acwt 10.0 \
-	  --nj $decode_nj --cmd "$decode_cmd" $iter_opts \
-	  --online-ivector-dir $test_ivector_dir \
+	  --nj $decode_nj --cmd "$decode_cmd" $iter_opts $decode_opt \
+      ${extra_left_context:+ --extra-left-context $extra_left_context} \
+      ${extra_right_context:+ --extra-right-context $extra_right_context} \
+      ${extra_left_context_initial:+ --extra-left-context-initial $extra_left_context_initial} \
+      ${extra_right_context_final:+ --extra-right-context-final $extra_right_context_final} \
+      ${frames_per_chunk_primary:+ --frames-per-chunk "$frames_per_chunk_primary"} \
+	  ${test_ivector_dir:+ --online-ivector-dir $test_ivector_dir} \
 	  $graph_dir data/${decode_set} \
 	  $dir/decode_${decode_set}${decode_iter:+_$decode_iter}_sw1_tg || exit 1;
 	  
 	  steps/lmrescore_const_arpa.sh --cmd "$decode_cmd" \
-		data/lang_sw1_{tg,fsh_fg} data/${decode_set} \
+		${LM_path}sw1_{tg,fsh_fg} data/${decode_set} \
 		$dir/decode_${decode_set}${decode_iter:+_$decode_iter}_sw1_{tg,fsh_fg} || exit 1;
 
   ) || touch $dir/.error &
